+++ app/Http/Controllers/ItemCategoryController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Route\Validators\Item as ItemRouteValidator;
 use App\Models\Category;
-use App\Models\Item;
 use App\Models\ItemCategory;
 use App\Transformers\ItemCategory as ItemCategoryTransformer;
 use App\Validators\ItemCategory as ItemCategoryValidator;
@@ -33,15 +33,15 @@ class ItemCategoryController extends Controller
      */
     public function index(Request $request, string $resource_type_id, string $resource_id, string $item_id): JsonResponse
     {
-        $item_category = (new ItemCategory())
-            ->where('item_id', '=', $item_id)
-            ->whereHas('item', function ($query) use ($resource_id, $resource_type_id) {
-                $query->where('resource_id', '=', $resource_id)
-                    ->whereHas('resource', function ($query) use ($resource_type_id) {
-                        $query->where('resource_type_id', '=', $resource_type_id);
-                    });
-            })
-            ->first();
+        if (ItemRouteValidator::validate($resource_type_id, $resource_id, $item_id) === false) {
+            return $this->returnResourceNotFound();
+        }
+
+        $item_category = (new ItemCategory())->paginatedCollection(
+            $resource_type_id,
+            $resource_id,
+            $item_id
+        );
 
         if ($item_category === null) {
             return $this->returnResourceNotFound();
@@ -77,15 +77,16 @@ class ItemCategoryController extends Controller
         string $item_category_id
     ): JsonResponse
     {
-        $item_category = (new ItemCategory())
-            ->where('item_id', '=', $item_id)
-            ->whereHas('item', function ($query) use ($resource_id, $resource_type_id) {
-                $query->where('resource_id', '=', $resource_id)
-                    ->whereHas('resource', function ($query) use ($resource_type_id) {
-                        $query->where('resource_type_id', '=', $resource_type_id);
-                    });
-            })
-            ->find($item_category_id);
+        if (ItemRouteValidator::validate($resource_type_id, $resource_id, $item_id) === false) {
+            return $this->returnResourceNotFound();
+        }
+
+        $item_category = (new ItemCategory())->single(
+            $resource_type_id,
+            $resource_id,
+            $item_id,
+            $item_category_id
+        );
 
         if ($item_category === null) {
             return $this->returnResourceNotFound();
@@ -114,14 +115,7 @@ class ItemCategoryController extends Controller
      */
     public function optionsIndex(Request $request, string $resource_type_id, string $resource_id, string $item_id): JsonResponse
     {
-        $item = (new Item())
-            ->where('resource_id', '=', $resource_id)
-            ->whereHas('resource', function ($query) use ($resource_type_id) {
-                $query->where('resource_type_id', '=', $resource_type_id);
-            })
-            ->find($item_id);
-
-        if ($item === null) {
+        if (ItemRouteValidator::validate($resource_type_id, $resource_id, $item_id) === false) {
             return $this->returnResourceNotFound();
         }
 
@@ -153,15 +147,16 @@ class ItemCategoryController extends Controller
         string $item_category_id
     ): JsonResponse
     {
-        $item_category = (new ItemCategory())
-            ->where('item_id', '=', $item_id)
-            ->whereHas('item', function ($query) use ($resource_id, $resource_type_id) {
-                $query->where('resource_id', '=', $resource_id)
-                    ->whereHas('resource', function ($query) use ($resource_type_id) {
-                        $query->where('resource_type_id', '=', $resource_type_id);
-                    });
-            })
-            ->find($item_category_id);
+        if (ItemRouteValidator::validate($resource_type_id, $resource_id, $item_id) === false) {
+            return $this->returnResourceNotFound();
+        }
+
+        $item_category = (new ItemCategory())->single(
+            $resource_type_id,
+            $resource_id,
+            $item_id,
+            $item_category_id
+        );
 
         if ($item_category === null) {
             return $this->returnResourceNotFound();
@@ -192,23 +187,8 @@ class ItemCategoryController extends Controller
         string $item_id
     ): JsonResponse
     {
-        $item = (new Item())
-            ->where('resource_id', '=', $resource_id)
-            ->whereHas('resource', function ($query) use ($resource_type_id) {
-                $query->where('resource_type_id', '=', $resource_type_id);
-            })
-            ->find($item_id);
-
-        if ($item === null) {
+        if (ItemRouteValidator::validate($resource_type_id, $resource_id, $item_id) === false) {
             return $this->returnResourceNotFound();
-        }
-
-        $item_category = (new ItemCategory())
-            ->where('item_id', '=', $item_id)
-            ->first();
-
-        if ($item_category !== null) {
-            return $this->returnResourceConflict();
         }
 
         $validator = (new ItemCategoryValidator)->create($request);
@@ -280,15 +260,16 @@ class ItemCategoryController extends Controller
         string $item_category_id
     ): JsonResponse
     {
-        $item_category = (new ItemCategory())
-            ->where('item_id', '=', $item_id)
-            ->whereHas('item', function ($query) use ($resource_id, $resource_type_id) {
-                $query->where('resource_id', '=', $resource_id)
-                    ->whereHas('resource', function ($query) use ($resource_type_id) {
-                        $query->where('resource_type_id', '=', $resource_type_id);
-                    });
-            })
-            ->find($item_category_id);
+        if (ItemRouteValidator::validate($resource_type_id, $resource_id, $item_id) === false) {
+            return $this->returnResourceNotFound();
+        }
+
+        $item_category = (new ItemCategory())->single(
+            $resource_type_id,
+            $resource_id,
+            $item_id,
+            $item_category_id
+        );
 
         if ($item_category === null) {
             return $this->returnResourceNotFound();